Modular testing. Software vs science test modules.
